+++ src/callbacks/project_page/callback_create_projects.py
@@ -4,8 +4,14 @@ from src.component_ids import MULTI_SELECT_SECTION_FOR_PROJECT_ID, EDITABLE_IMPO
     STORED_IMPORTED_RUNS_DATA, \
     TABLE_PROJECT_SUMMARY_ID, ADD_PROJECT_BUTTON_ID, PROJECT_NAME_INPUT_FIELD_ID, ALERT_PROJECT_CREATION_ID, \
     STORED_PROJECT_OVERVIEW_DATA, UPDATE_PROJECT_BUTTON_ID, PROJECT_START_YEAR_INPUT_FIELD_ID, \
-    PROJECT_END_YEAR_INPUT_FIELD_ID
+    PROJECT_END_YEAR_INPUT_FIELD_ID, OVERVIEW_PROJECT_MAP_ID_2, PROGRAM_SELECTION_MAP_RADIO_SWITCH_ID
+from src.constants import ProgramDefinitionMapType
 from src.linear_objects.dike_traject import DikeTraject
+from src.linear_objects.project import get_projects_from_saved_data
+from src.plotly_graphs.plotly_maps import plot_default_overview_map_dummy
+from src.plotly_graphs.project_page.plotly_maps import plot_project_overview_map, \
+    plot_comparison_runs_overview_map_simple, plot_comparison_runs_overview_map_projects, \
+    plot_comparison_runs_overview_map_assessment
 
 
 @callback(
@@ -232,3 +238,32 @@ def update_section_selection_on_click_event(selected_row: dict, project_data_ove
             selected_sections = project['sections']
             return selected_sections, selected_project_name
     return dash.no_update, dash.no_update
+
+
+@callback(
+    Output(OVERVIEW_PROJECT_MAP_ID_2, "figure"),
+    [Input("tabs_tab_project_page", "value"),
+     # Input(TABLE_PROJECT_SUMMARY_ID, "selectedRows"),
+     Input(MULTI_SELECT_SECTION_FOR_PROJECT_ID, "value"),
+     Input(PROGRAM_SELECTION_MAP_RADIO_SWITCH_ID, "value"),
+     State(STORED_IMPORTED_RUNS_DATA, "data"),
+     State(STORED_PROJECT_OVERVIEW_DATA, "data")]
+)
+def update_map_project_definition_page(dummy, selected_sections: list, switch_type_map: str, imported_runs_data: dict,
+                                       project_overview_data: list):
+    if imported_runs_data is None:
+        return dash.no_update
+    if project_overview_data is None:
+        return dash.no_update
+
+    projects, trajects = get_projects_from_saved_data(imported_runs_data, project_overview_data, calc_failure_pro=False)
+
+    if switch_type_map == ProgramDefinitionMapType.SIMPLE.name:
+        _fig = plot_comparison_runs_overview_map_simple(list(trajects.values()), selected_sections)
+    elif switch_type_map == ProgramDefinitionMapType.PROJECTS.name:
+        _fig = plot_comparison_runs_overview_map_projects(projects, list(trajects.values()))
+    elif switch_type_map == ProgramDefinitionMapType.ASSESSMENT_PROBABILITIES.name:
+        _fig = plot_comparison_runs_overview_map_assessment(list(trajects.values()))
+    else:
+        _fig = plot_default_overview_map_dummy()
+    return _fig