+++ src/callbacks/project_page/callback_import_run.py
@@ -7,7 +7,6 @@ from src.component_ids import STORED_IMPORTED_RUNS_DATA, EDITABLE_IMPORTED_RUNS_
 import base64
 import json
 
-from src.plotly_graphs.project_page.plotly_maps import plot_comparison_runs_overview_map
 
 
 @callback(
@@ -53,11 +52,11 @@ def upload_and_save_in_project_data(contents: str, filename: str, stored_importe
 #
 @callback(
     Output(EDITABLE_IMPORTED_RUNS_TABLE_ID, "rowData"),
-    Output(OVERVIEW_PROJECT_MAP_ID_2, "figure"),
+    # Output(OVERVIEW_PROJECT_MAP_ID_2, "figure"),
     Input(STORED_IMPORTED_RUNS_DATA, "data"),
     Input("tabs_tab_project_page", "active_tab")
 )
-def fill_table_project_overview_and_update_map(imported_runs_data: dict, dummy: str) -> tuple[list[dict], go.Figure]:
+def fill_table_project_overview_and_update_map(imported_runs_data: dict, dummy: str) -> list[dict]:
     """
     Fill the overview table with the project data wth the imported dike traject data.
     :param project_data:
@@ -68,14 +67,14 @@ def fill_table_project_overview_and_update_map(imported_runs_data: dict, dummy: 
     row_data = []
 
     if imported_runs_data is None:
-        return dash.no_update, dash.no_update
+        return dash.no_update
     if imported_runs_data == {}:
-        return dash.no_update, dash.no_update
+        return dash.no_update
 
     for traject_run in imported_runs_data.keys():
         traject = traject_run
         run = imported_runs_data[traject_run]["run_name"]
         row_data.append({"traject": traject, "run_name": run, "active": False})
-    _fig = plot_comparison_runs_overview_map(imported_runs_data)
+    # _fig = plot_comparison_runs_overview_map(imported_runs_data)
 
-    return row_data, _fig
+    return row_data