+++ src/layouts/layout_project_page/layout_project_definition_tab.py
@@ -9,8 +9,9 @@ import dash_mantine_components as dmc
 from src.component_ids import EDITABLE_IMPORTED_RUNS_TABLE_ID, TABLE_PROJECT_SUMMARY_ID, \
     MULTI_SELECT_SECTION_FOR_PROJECT_ID, ADD_PROJECT_BUTTON_ID, PROJECT_NAME_INPUT_FIELD_ID, ALERT_PROJECT_CREATION_ID, \
     UPDATE_PROJECT_BUTTON_ID, PROJECT_START_YEAR_INPUT_FIELD_ID, DELETE_PROJECT_BUTTON_ID, \
-    PROJECT_END_YEAR_INPUT_FIELD_ID, EXPORT_PROJECTS_TO_JSON_ID, BUTTON_DOWNLOAD_PROJECTS_EXPORT, UPLOAD_SAVED_PROJECTS, \
-    CONTENT_TABS_COMPARISON_PAGE_ID, OVERVIEW_COMPARISON_MAP_ID, OVERVIEW_PROJECT_MAP_ID_2
+    PROJECT_END_YEAR_INPUT_FIELD_ID, UPLOAD_SAVED_PROJECTS, OVERVIEW_PROJECT_MAP_ID_2, \
+    PROGRAM_SELECTION_MAP_RADIO_SWITCH_ID
+from src.constants import ProgramDefinitionMapType
 from src.layouts.layout_traject_page.layout_download_buttons import layout_download_projects
 from src.plotly_graphs.plotly_maps import plot_default_overview_map_dummy
 
@@ -95,6 +96,17 @@ multi_select = dmc.MultiSelect(
     w=400,
     mb=10,
     clearable=True,
+)
+
+layout_radio_helper_map_switch = dbc.RadioItems(
+    id=PROGRAM_SELECTION_MAP_RADIO_SWITCH_ID,
+    options=[
+        {"label": ProgramDefinitionMapType.SIMPLE.value, "value": ProgramDefinitionMapType.SIMPLE.name},
+        {"label": ProgramDefinitionMapType.PROJECTS.value, "value": ProgramDefinitionMapType.PROJECTS.name},
+        {"label": ProgramDefinitionMapType.ASSESSMENT_PROBABILITIES.value, "value": ProgramDefinitionMapType.ASSESSMENT_PROBABILITIES.name},
+    ],
+    value="SIMPLE",
+    inline=True
 )
 
 left_side = [
@@ -185,12 +197,15 @@ right_side = [
         children=[
 
             html.Div(
-                style={'width': '90vh', 'height': '60vh', 'border': "2px solid black"},
-                children=dcc.Graph(
-                    id=OVERVIEW_PROJECT_MAP_ID_2,
-                    figure=plot_default_overview_map_dummy(),
-                    style={"width": "100%", "height": "100%"},
-                )
+                style={'width': '110vh', 'height': '80vh', 'border': "2px solid black"},
+                children=[
+                    layout_radio_helper_map_switch,
+                    dcc.Graph(
+                        id=OVERVIEW_PROJECT_MAP_ID_2,
+                        figure=plot_default_overview_map_dummy(),
+                        style={"width": "100%", "height": "100%"},
+                    )
+                ]
             ),
 
         ])