+++ src/plotly_graphs/project_page/plotly_maps.py
@@ -3,12 +3,14 @@ from typing import Optional
 import numpy as np
 import plotly.graph_objects as go
 
-from src.constants import PROJECTS_COLOR_SEQUENCE
+from src.constants import PROJECTS_COLOR_SEQUENCE, Mechanism, ResultType, ColorBarResultType, SubResultType
 from src.linear_objects.dike_traject import DikeTraject
 from src.linear_objects.project import DikeProject
 from src.plotly_graphs.plotly_maps import update_layout_map_box, add_section_trace, plot_default_overview_map_dummy, \
-    get_middle_point, get_average_point
+    get_middle_point, get_average_point, get_reliability_color, add_colorscale_bar, place_legend_left_top_corner, \
+    place_legend_right_top_corner
 from src.utils.gws_convertor import GWSRDConvertor
+from src.utils.utils import get_beta, beta_to_pf
 
 
 def plot_project_overview_map(projects: list[DikeProject], trajects: Optional[list[DikeTraject]]=None) -> go.Figure:
@@ -92,35 +94,60 @@ def plot_project_overview_map(projects: list[DikeProject], trajects: Optional[li
                         showlegend=True if index == 0 else False,
                     )
                 )
-
+    place_legend_right_top_corner(fig)
     return fig
 
+def plot_comparison_runs_overview_map_projects(projects: list[DikeProject], trajects: list[DikeTraject])-> go.Figure:
+    return plot_project_overview_map(projects, trajects)
 
-def plot_comparison_runs_overview_map(project_data: dict) -> go.Figure:
-    """
-    This function plots an overview Map of the current dike in data. It uses plotly Mapbox for the visualization.
+def plot_comparison_runs_overview_map_assessment(trajects: list[DikeTraject])-> go.Figure:
 
-    :param dike_traject: DikeTraject object with the data of the dike.
-    :param selected_result_type: string of the selected result type in the select dropdown field.
-
-    :return:
-    """
     fig = go.Figure()
-    traject_plotted = []
-    for _, dike_traject_data in project_data.items():
-        dike_traject = DikeTraject.deserialize(dike_traject_data)
-        if dike_traject.name in traject_plotted:
-            continue
+    sections = []  # add section to a list to find the middle point for all trajects
+    for dike_traject in trajects:
+        for section in dike_traject.dike_sections:
+            sections.append(section)
+            _coordinates_wgs = [
+                GWSRDConvertor().to_wgs(pt[0], pt[1]) for pt in section.coordinates_rd
+            ]  # convert in GWS coordinates:
 
-        # pick a random color for the dike traject
-        _color = f"rgb({np.random.randint(0, 255)}, {np.random.randint(0, 255)}, {np.random.randint(0, 255)})"
-        showlegend = True
-        for index, section in enumerate(dike_traject.dike_sections):
             # if a section is not in analyse, skip it, and it turns blank on the map.
-            _hovertemplate = (
-                    f"Traject {dike_traject.name}<br>" +
-                    f"Vaknaam {section.name}<br>" + f"Lengte: {section.length}m <extra></extra>"
-            )
+            if not section.in_analyse:
+                continue
+
+            _initial_results = section.initial_assessment
+
+            if _initial_results is not None:
+                _year_index = 0
+                _beta = get_beta(_initial_results, _year_index, Mechanism.SECTION.name)
+                _beta_dict = {
+                    meca: beta[_year_index]
+                    for meca, beta in _initial_results.items()
+                    if meca != "Section"
+                }
+                _color = get_reliability_color(_beta, dike_traject.lower_bound_value)
+
+
+                _hover_res = f"Pf sectie: {beta_to_pf(_beta):.2e}<br>"
+
+                _hovertemplate = (
+                        f"Vaknaam {section.name}<br>" + _hover_res + "<extra></extra>"
+                )
+
+                _mechanism = min(
+                    _beta_dict, key=_beta_dict.get
+                )  # mechanism with lowest beta
+                _hovertemplate = (
+                        _hovertemplate[:-15]
+                        + f"Laagste beta: {_mechanism}<br>"
+                        + "<extra></extra>"
+                )  # :-15 to remove <extra></extra> from string
+
+            else:
+                _color = "grey"
+                _hovertemplate = (
+                        f"Vaknaam {section.name}<br>" f"Beta: NO DATA<br>" + "<extra></extra>"
+                )
 
             add_section_trace(
                 fig,
@@ -128,26 +155,70 @@ def plot_comparison_runs_overview_map(project_data: dict) -> go.Figure:
                 name=dike_traject.name,
                 color=_color,
                 hovertemplate=_hovertemplate,
-                showlegend=showlegend,
-                legendgroup=dike_traject.name
             )
-            showlegend = False
-            traject_plotted.append(dike_traject.name)
+
+    # Add colorscale bar, /!\ This will be centered around the lower bound value of the LAST dike traject
+    add_colorscale_bar(
+        fig,
+        ResultType.PROBABILITY.name,
+        ColorBarResultType.RELIABILITY.name,
+        SubResultType.ABSOLUTE.name,
+        dike_traject.lower_bound_value,
+    )
 
     # Update layout of the figure and add token for mapbox
-    _middle_point = (52.155170, 5.387207)  # lat/lon of Amersfoort
-    update_layout_map_box(fig, _middle_point, zoom=7)
+    _middle_point = get_middle_point(sections)
+    update_layout_map_box(fig, _middle_point)
+    place_legend_right_top_corner(fig)
 
-    # move legend to the left
-    fig.update_layout(
-        legend=dict(
-            orientation="h",
-            yanchor="bottom",
-            y=1.02,
-            xanchor="left",
-            x=0.01,
-        )
-    )
+    return fig
+
+def plot_comparison_runs_overview_map_simple(trajects: list[DikeTraject], selected_sections)-> go.Figure:
+    """
+    This function plots an overview Map of the current dike in data. It uses plotly Mapbox for the visualization.
+
+
+    :return:
+    """
+    fig = go.Figure()
+    sections = []
+
+    if trajects is not None:
+        for traject in trajects:
+            for index, section in enumerate(traject.dike_sections):
+                sections.append(section)
+
+                if f"{section.name}|{traject.name}" in selected_sections:
+                    _color = "red"
+                else:
+                    _color = "grey"
+                # if a section is not in analyse, skip it, and it turns blank on the map.
+                _hovertemplate = (
+                        f"Traject {traject.name}<br>" +
+                        f"Vaknaam {section.name}<br>" + f"Lengte: {section.length}m <extra></extra>"
+                )
+                _coordinates_wgs = [
+                    GWSRDConvertor().to_wgs(pt[0], pt[1]) for pt in section.coordinates_rd
+                ]  # convert in GWS coordinates:
+
+                fig.add_trace(
+                    go.Scattermap(
+                        mode="lines",
+                        lat=[x[0] for x in _coordinates_wgs],
+                        lon=[x[1] for x in _coordinates_wgs],
+                        marker={"size": 10, "color": _color},
+                        line={"width": 4, "color": _color},
+                        name=traject.name,
+                        legendgroup=traject.name,
+                        hovertemplate=_hovertemplate,
+                        opacity=0.9,
+                        showlegend=True if index == 0 else False,
+                    )
+                )
+
+    _middle_point = get_average_point(sections)
+    update_layout_map_box(fig, _middle_point, zoom=10)
+    place_legend_right_top_corner(fig)
 
     return fig