+++ tests/test_plotly_figures/test_plotly_scattermapbox_figures_traject_page.py
@@ -8,8 +8,9 @@ from plotly.graph_objs import Figure
 from src.constants import ResultType, Mechanism, SubResultType, ColorBarResultType, CalcType
 from src.linear_objects.dike_traject import DikeTraject
 from src.plotly_graphs.plotly_maps import plot_dike_traject_reliability_initial_assessment_map, \
-    plot_dike_traject_reliability_measures_assessment_map, plot_dike_traject_urgency, dike_traject_pf_cost_helping_map, \
-    plot_default_overview_map_dummy, plot_overview_map
+    plot_dike_traject_reliability_measures_assessment_map, plot_dike_traject_urgency, \
+    dike_traject_pf_cost_helping_map_simple, \
+    plot_default_overview_map_dummy, plot_overview_map, dike_traject_pf_cost_helping_map_detail
 
 
 class TestPlotlyScatterMapBox():
@@ -32,7 +33,6 @@ class TestPlotlyScatterMapBox():
 
         # 3. Assert
         assert isinstance(_fig, Figure)
-
 
     @pytest.mark.parametrize("result_type", [ResultType.RELIABILITY, ResultType.PROBABILITY])
     @pytest.mark.parametrize("mechanism",
@@ -96,9 +96,11 @@ class TestPlotlyScatterMapBox():
         assert isinstance(_fig, Figure)
 
     @pytest.mark.parametrize("calc_type", [CalcType.VEILIGHEIDSRENDEMENT, CalcType.DOORSNEDE_EISEN])
-    @pytest.mark.parametrize("sub_result_type", [SubResultType.MEASURE_TYPE, SubResultType.BERM_WIDENING, SubResultType.CREST_HIGHTENING, SubResultType.INVESTMENT_YEAR])
-    def test_plot_dike_traject_reliability_measures_assessment_map_measure(self,calc_type: Enum,
-                                                                        sub_result_type: Enum):
+    @pytest.mark.parametrize("sub_result_type",
+                             [SubResultType.MEASURE_TYPE, SubResultType.BERM_WIDENING, SubResultType.CREST_HIGHTENING,
+                              SubResultType.INVESTMENT_YEAR])
+    def test_plot_dike_traject_reliability_measures_assessment_map_measure(self, calc_type: Enum,
+                                                                           sub_result_type: Enum):
         # 1. Define data
         _dike_data = json.load(
             open(Path(__file__).parent.parent / 'data/31-1 base coastal case/reference' / 'dike_data.json'))
@@ -129,16 +131,31 @@ class TestPlotlyScatterMapBox():
         # 3. Assert
         assert isinstance(_fig, Figure)
 
-    def test_dike_traject_pf_cost_helping_map(self):
+    def test_dike_traject_pf_cost_helping_map_simple(self):
         # 1. Define data
         _dike_data = json.load(
             open(Path(__file__).parent.parent / 'data/31-1 base coastal case/reference' / 'dike_data.json'))
         _dike_traject = DikeTraject.deserialize(_dike_data)
 
         # 2. Call
-        _fig1 = dike_traject_pf_cost_helping_map(_dike_traject, 0, ["33"])
-        _fig2 = dike_traject_pf_cost_helping_map(_dike_traject, 0, ["93"])  # this one has a grey dijkvak
+        _fig1 = dike_traject_pf_cost_helping_map_simple(_dike_traject, 0, ["33"])
+        _fig2 = dike_traject_pf_cost_helping_map_simple(_dike_traject, 0, ["93"])  # this one has a grey dijkvak
 
         # 3. Assert
         assert isinstance(_fig1, Figure)
         assert isinstance(_fig2, Figure)
+
+    def test_dike_traject_pf_cost_helping_map_detailed(self):
+        # 1. Define data
+        _dike_data = json.load(
+            open(Path(__file__).parent.parent / 'data/31-1 base coastal case/reference' / 'dike_data.json'))
+        _dike_traject = DikeTraject.deserialize(_dike_data)
+
+        # 2. Call
+        _fig1 = dike_traject_pf_cost_helping_map_detail(_dike_traject, 0,
+                                                        [
+                                                            "WsNoo_Stab_014700_015400",
+                                                            "WsNoo_Stab_014100_014700",
+                                                            "WsNoo_Stab_017400_018000"])
+        # 3. Assert
+        assert isinstance(_fig1, Figure)