+++ docs/CHANGELOG.md
@@ -3,10 +3,13 @@
 ## v0.0.0 [##/##/####]
 
 ### Added
-None.
+- Added a button and name field to export the traject run json
+- Added a switch for the map of the program definition to display faalkans
+- Added a comparison map for the measures of 2 runs
 
 ### Changed
-None.
+- Map below reliability vs cost graph now displays the selected measure
+- Map of the program definition now displays the sections selected in the MultiSelectField
 
 ### Deprecated
 None.